backend/requirements.txt
alembic==1.16.2
argcomplete==3.6.2
blinker==1.9.0
boto3==1.20.24
botocore==1.23.24
certifi==2025.6.15
cfn-flip==1.3.0
charset-normalizer==3.4.2
click==8.1.8
durationpy==0.10
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
hjson==3.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
kappa==0.6.0
Mako==1.3.10
MarkupSafe==3.0.2
placebo==0.9.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.5.2
serverless-wsgi==3.1.0
six==1.17.0
SQLAlchemy==2.0.41
text-unidecode==1.3
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
troposphere==4.9.3
typing_extensions==4.14.0
urllib3==1.26.20
Werkzeug==3.1.3
zappa==0.60.2
zipp==3.23.0
